\name{internal.functions}
\alias{ck.colors}
\alias{f.build.pixel}
\alias{f.points.CKrige}
\alias{f.polygons.CKrige}
\alias{f.covmat.support}
\alias{f.point.block.cov}
\alias{f.point.rec.covmat}
\alias{f.kriging.method}
\alias{f.point.block.cov}
\alias{f.uk.pred}
\alias{f.ck.pred}
\alias{f.cmck.pred}
\alias{f.points.preCKrige}
\alias{f.polygons.preCKrige}
\alias{f.polygoncovmat}
\alias{f.bb.cov.one.configuration}
\alias{f.bbox.information}
\alias{f.build.pixel}
\alias{f.calc.P}
\alias{f.calc.Q}
\alias{f.centroid.polygon}
\alias{f.gpc.poly}
\alias{f.indicate.small.pred.locations}
\alias{f.intersect.area}
\alias{f.intersect.max.area}
\alias{f.p.in.poly}
\alias{f.pixconfig}
\alias{f.pixelcovmat}
\alias{f.pixelgrid}
\alias{f.pp.cov}
\alias{f.row.dist}
\alias{f.preCKrige.check.and.test}
\alias{K}
\alias{gpc.copyright} 
%- Also NEED an '\alias' for EACH other topic documented here.
\title{Internal functions}
\description{
To show the R-code of the above listed internal functions use 
\code{constrainedKriging:::function name}
}
\author{Christoph Hofer}
\keyword{internal}

