\name{preCKrigePoints-class}
\Rdversion{1.1}
\docType{class}
\alias{preCKrigePoints-class}
\alias{print.preCKrigePoints}
\alias{summary.preCKrigePoints}
\alias{show,preCKrigePoints-method}


\title{Class "preCKrigePoints"}
\description{Class of objects that are generated by \code{\link{preCKrige}} if the attribute \code{newdata} is of the class \cr
\code{SpatialPoints} or \code{SpatialPointsDataFrame}. This class has a \code{show} and \code{summary} method.}
\section{Objects from the Class}{
Objects can be created by calls of the form \code{new("preCKrigePoints", ...)}.
}
\section{Slots}{
  \describe{
    \item{\code{covmat}:}{Object of class \code{"list"} }
    \item{\code{posindex}:}{Object of class \code{"list"} }
    \item{\code{model}:}{Object of class \code{"list"}  }
    \item{\code{data}:}{Object of class \code{"data.frame"}  }
    \item{\code{coords}:}{Object of class \code{"matrix"}  }
  }
}
\section{Methods}{
  \describe{
    \item{CKrige}{\code{signature(formula = "formula", data = "data.frame", locations = "formula", object = "preCKrigePoints", type = "numeric", ex.out = "logical")}: ... }
	 }
}
\author{Christoph Hofer \email{christoph.hofer@alumni.ethz.ch}}
\seealso{\code{\link{preCKrige}}, 
\code{\linkS4class{preCKrigePolygons}}  }
\examples{
showClass("preCKrigePoints")
}
\keyword{classes}
