\name{preCKrigePolygons-class}
\Rdversion{1.1}
\docType{class}
\alias{preCKrigePolygons-class}
\alias{print.preCKrigePolygons}
\alias{summary.preCKrigePolygons}
\alias{show-methods}
\alias{show,preCKrigePolygons-method}

\title{Class "preCKrigePolygons" ~~~ }
\description{Class of objects that are generated by \code{\link{preCKrige}} if the attribute \code{newdata} is of the class \cr
\code{SpatialPolygons} or \code{SpatialPolygonsDataFrame}. This class has a \code{show} and \code{summary} method.}
\section{Objects from the Class}{
Objects can be created by calls of the form \code{new("preCKrigePolygons", ...)}.
}
\section{Slots}{
  \describe{
    \item{\code{covmat}:}{Object of class \code{"list"}}
    \item{\code{se.covmat}:}{Object of class \code{"list"}}
    \item{\code{pixconfig}:}{Object of class \code{"list"} }
    \item{\code{pixcovmat}:}{Object of class \code{"matrix"}}
    \item{\code{model}:}{Object of class \code{"list"} ~~ }
    \item{\code{data}:}{Object of class \code{"data.frame"}}
    \item{\code{polygons}:}{Object of class \code{"list"}}
  }
}
\section{Methods}{
  \describe{
    \item{CKrige}{\code{signature(formula = "formula", data = "data.frame", locations = "formula", object = "preCKrigePolygons", type = "numeric", ex.out = "logical")}: ... }
	 }
}
\references{--}
\author{Christoph Hofer}


\seealso{\code{\link{preCKrige}}, \code{\linkS4class{preCKrigePoints}}
}
\examples{
showClass("preCKrigePolygons")
}
\keyword{classes}
